Fixed # 33162: Android Visit notes given by user is not showing up on WEB Providers intake screen
Details:
a. Hardcoded the visit cost temporarily as the cost should go into new screen on its own
b. Visit object needs to be created after taking the reason_for_visit which is on NEXT click
c. if we donot want to hardcode cost, and rather fetch it from visit object, then visit needs to be created on load of this screen; and that is not taking the reason for visit once it gets created and that is the Cause of this bug.